backend/requirements.txt
djongo==1.3.6
pymongo==3.12.3
django
djangorestframework
